feat: 요금정책 비지니스로직 서비스 계층으로 분리

+++ src/test/java/nextstep/subway/acceptance/PathAcceptanceTest.java
@@ -156,7 +156,7 @@ class PathAcceptanceTest extends AcceptanceTest {
                 () -> assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(남부터미널역, 교대역, 강남역),
                 () -> assertThat(response.jsonPath().getInt("distance")).isEqualTo(12),
                 () -> assertThat(response.jsonPath().getInt("duration")).isEqualTo(12),
-                () -> assertThat(response.jsonPath().getInt("fare")).isEqualTo((1_250 + 100 + 900) * 0.8)
+                () -> assertThat(response.jsonPath().getInt("fare")).isEqualTo((int) ((1_250 + 100 + 900) * 0.8))
         );
     }
 
@@ -180,7 +180,7 @@ class PathAcceptanceTest extends AcceptanceTest {
                 () -> assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(남부터미널역, 교대역, 강남역),
                 () -> assertThat(response.jsonPath().getInt("distance")).isEqualTo(12),
                 () -> assertThat(response.jsonPath().getInt("duration")).isEqualTo(12),
-                () -> assertThat(response.jsonPath().getInt("fare")).isEqualTo((1_250 + 100 + 900) * 0.5)
+                () -> assertThat(response.jsonPath().getInt("fare")).isEqualTo((int) ((1_250 + 100 + 900) * 0.5))
         );
     }
